Units convert

+++ app/src/main/java/com/personal/weathering/domain/models/weather/CurrentWeatherData.kt
@@ -4,8 +4,8 @@ import java.time.LocalDateTime
 
 data class CurrentWeatherData(
     val time: LocalDateTime,
-    val temperature: Int,
-    val apparentTemperature: Int,
+    val temperature: Double,
+    val apparentTemperature: Double,
     val humidity: Int,
     val humidityType: HumidityType,
     val weatherType: WeatherType,
